# By holtgrew (232) and others
# Via holtgrew
* master: (383 commits)
  [FIX] Removing RemoveOuterConst again together with removing inner const of pointers.
  [FIX] Fixing roi_feature_projection, BED format detection.
  [FIX,FEATURE] Adding RemoveOuterConst and using it in allocation code.
  [DOC,INTERNAL] Adding note.
  [DOC] Correcting parameter name in documentation of ArgParseOption.
  [FIX] Removing bad part of previous commit.
  [DOC] Adding demos from DDDoc to dox, using @include feature for stdout.
  [DOC] Dox documentation for the text position manipulation functions.
  [FIX,TEST] Fixing bug in mason_variator, adding test for this case.
  [FIX] Fixing const issue for integer concept.
  [FIX] Fixed bug with banded globalAlignmentScore() function. Works now for segments too.
  [FIX] extendAlignment() with x-Drop value.
  [FIX] Fixing BamStream problem in MSVC
  [FEATURE] more readable error message in case of differing text files
  [FIX] several bugs fixed: endless loop for GFF lines without key-value pair, missing quotation of key-value separator, quote all non-integral values in GTF
  [DOC] Adding do Mason Simulator documentation.
  [FIX] Fixing methylation computation plus methylation seeds in Mason2.
  [DOC] Demo for alignExtend().
  [FIX,INTERNAL,DOC] fix minor warnings, errors and doc
  [DOC] Improvded documentation and demo for extendAlignment.
  ...

Conflicts:
	core/include/seqan/index/index_fm.h
	core/include/seqan/index/index_fm_prefix_sum_table.h
	core/include/seqan/index/index_fm_rank_dictionary_bms.h
	core/include/seqan/index/index_fm_rank_support_bit_string.h
	core/include/seqan/index/index_fm_rank_support_bit_string_iterator.h
	core/include/seqan/index/index_fm_sentinel_rank_dictionary.h
	core/include/seqan/platform.h
	core/include/seqan/platform/platform_gcc.h
	extras/apps/masai/store.h
	util/cmake/SeqAnLegacyBuildSystem.cmake

+++ core/include/seqan/index/index_fm_prefix_sum_table.h
@@ -33,6 +33,8 @@
 // Author: Jochen Singer <jochen.singer@fu-berlin.de>
 // ==========================================================================
 
+//SEQAN_NO_DDDOC:do not generate documentation for this file
+
 #ifndef INDEX_FM_PREFIX_SUM_TABLE_H_
 #define INDEX_FM_PREFIX_SUM_TABLE_H_
 
@@ -94,12 +96,13 @@ struct GetValue<PrefixSumTable<TChar, TSpec> const>
 // ----------------------------------------------------------------------------
 // Metafunction Fibre
 // ----------------------------------------------------------------------------
+// TODO(singer): change the type to be dependent on a template argument
 // TODO(esiragusa): It should be possible to deduce the text size.
 
 template <typename TChar, typename TSpec>
 struct Fibre<PrefixSumTable<TChar, TSpec>, FibreEntries>
 {
-    typedef String<unsigned> Type;
+    typedef String<unsigned long long> Type;
 };
 
 // ----------------------------------------------------------------------------
